+++ signbank/dictionary/forms.py
@@ -292,6 +292,16 @@ class MorphologyForm(forms.ModelForm):
         fields = ['role']
 
 
+class CSVFileOnlyUpload(forms.Form):
+    file = forms.FileField()
+
+    def clean_file(self):
+        file = self.cleaned_data['file']
+        if not file.name.endswith('.csv'):
+            raise forms.ValidationError(_('Must be a CSV file with .csv extension.'))
+        return file
+
+
 class CSVUploadForm(forms.Form):
     file = forms.FileField()
     dataset = forms.ModelChoiceField(queryset=Dataset.objects.all(), empty_label=None)
