+++ signbank/dictionary/templates/dictionary/import_qualtrics_csv.html
@@ -1,0 +1,33 @@
+{% extends 'baselayout.html' %}
+{% load stylesheet %}
+{% load bootstrap3 %}
+{% load i18n %}
+{% block bootstrap3_title %}{% blocktrans %}Import Qualtrics CSV{% endblocktrans %} | {% endblock %}
+
+{% block content %}
+    {% if perms.dictionary.import_csv %}
+    <div id="import-csv">
+        {# Translators: #}
+        <h3>{% blocktrans %}Import Validation Records from Qualtrics{% endblocktrans %}:</h3>
+
+        <ul>
+            <li>{% blocktrans %}CSV-file should be exported from Qualtrics and uploaded without modifications{% endblocktrans %}</li>
+            <li>{% blocktrans %}Import process expects responses to Qualtrics survey to be given in text format, not numerical format{% endblocktrans %}</li>
+            <li>{% blocktrans %}The first three rows of the file will not be added because it is expected to be the column
+                headers.{% endblocktrans %}</li>
+            <li>{% blocktrans %}The CSV-file's character set needs to be UTF-8. When exporting the CSV-file,
+                make sure that you select UTF-8 charset.{% endblocktrans %}</li>
+        </ul>
+
+        <form enctype="multipart/form-data" action='{% url "dictionary:import_qualtrics_csv" %}' method='post'>
+            {% csrf_token %}
+            {% bootstrap_field import_csv_form.file %}
+            <br>
+            <input class='btn btn-primary' type='submit' value='{% blocktrans %}Import validation records from CSV{% endblocktrans %}'>
+        </form>
+    </div>
+    {% else %}
+    {# Translators: Message that appears if user doesn't have proper user rights to view this page. #}
+    <p>{% blocktrans %}You do not have sufficient user rights to view this page.{% endblocktrans %}</p>
+    {% endif %}
+{% endblock %}
